refactor: 검색 결과 요청 파라미터 수정
