+++ frontend/src/hooks/queries/search/index.ts
@@ -1,1 +1,1 @@
-export { default as useInfiniteSearchedProductsQuery } from './useInfiniteSearchedProductsQuery';
+export { default as useInfiniteProductSearchResultsQuery } from './useInfiniteProductSearchResultsQuery';
