+++ frontend/src/hooks/queries/search/useInfiniteProductSearchResultsQuery.ts
@@ -4,15 +4,15 @@ import { searchApi } from '@/apis';
 import type { SearchedProductResponse } from '@/types/response';
 
 const fetchSearchedProducts = async (query: string, page: number) => {
-  const response = await searchApi.get({ params: '/products', queries: `?query=${query}&page=${page}` });
+  const response = await searchApi.get({ params: '/products/results', queries: `?query=${query}&page=${page}` });
   const data: SearchedProductResponse = await response.json();
 
   return data;
 };
 
-const useInfiniteSearchedProductsQuery = (query: string) => {
+const useInfiniteProductSearchResultsQuery = (query: string) => {
   return useInfiniteQuery({
-    queryKey: ['search', 'products', query],
+    queryKey: ['search', 'products', 'result', query],
     queryFn: ({ pageParam = 0 }) => fetchSearchedProducts(query, pageParam),
     getNextPageParam: (prevResponse: SearchedProductResponse) => {
       const isLast = prevResponse.page.lastPage;
@@ -23,4 +23,4 @@ const useInfiniteSearchedProductsQuery = (query: string) => {
   });
 };
 
-export default useInfiniteSearchedProductsQuery;
+export default useInfiniteProductSearchResultsQuery;
