+++ frontend/src/mocks/data/productSearchResults.json
